requirements.txt
asgiref==3.11.0
Django==4.2.27
django-admin-interface==0.32.0
django-ckeditor==6.7.3
django-colorfield==0.14.0
django-js-asset==2.2.0
django-sslserver==0.22
django-widget-tweaks==1.5.0
pillow==11.3.0
python-slugify==8.0.4
sqlparse==0.5.5
text-unidecode==1.3
typing_extensions==4.15.0
Unidecode==1.4.0
